Support IK on arduino

+++ arduino/common/CRC16.h
@@ -11,7 +11,7 @@ class CRC16
 		CRC16();
 
 		uint16_t CRCprocessByte(uint8_t data);
-		uint16_t CRCprocessBuffer(const uint8_t *data_p, uint16_t length);
+		uint16_t CRCprocessBuffer(const uint8_t *data_p, int length);
 		bool 	 CRCcheck(uint8_t *data_p, uint16_t length, uint16_t crc);
 
 	private:
